# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Alba Digital API",
        "description": "An API of ecommerce",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/categories/all": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "all categories",
                "description": "list of all categories",
                "operationId": "allCategories",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "category 1"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-08-28 06:06:17"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-08-28 06:06:17"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Paginated category list",
                "description": "Paginated list of categories",
                "operationId": "categoryList",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "category 1"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-08-28 06:06:17"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-08-28 06:06:17"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/category": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Create Category",
                "description": "Create a category record",
                "operationId": "categoryCreate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "category 1"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Product 1"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2023-08-28 06:06:17"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2023-08-28 06:06:17"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid data provided"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Invalid data provided"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}